✨ certif: use PixTable on team list
