--- certif/app/components/members-list-item.hbs
@@ -1,91 +1,0 @@
-<tr aria-label="Membres du centre de certification">
-  <td>{{@member.lastName}}</td>
-  <td>{{@member.firstName}}</td>
-  <td>
-    {{#if this.isEditionMode}}
-      <PixSelect
-        @screenReaderOnly={{true}}
-        @hideDefaultOption={{true}}
-        @placeholder="{{t 'pages.team.members.actions.select-role.label'}}"
-        @onChange={{this.setRoleSelection}}
-        @options={{this.roleOptions}}
-        @value={{@member.role}}
-      >
-        <:label>{{t "pages.team.members.actions.select-role.label"}}</:label>
-      </PixSelect>
-    {{else}}
-      {{@member.roleLabel}}
-    {{/if}}
-  </td>
-  {{#if this.shouldDisplayManagingColumn}}
-    <td>
-      {{#if this.shouldDisplayMemberManageButton}}
-        <Dropdown::IconTrigger
-          @icon="moreVert"
-          @dropdownButtonClass="zone-edit-role__dropdown-button"
-          @dropdownContentClass="managing-member__dropdown-content"
-          @ariaLabel={{t "pages.team.members.actions.manage"}}
-        >
-          {{#if this.shouldDisplayChangeRoleOption}}
-            <Dropdown::Item @onClick={{this.toggleEditionMode}}>
-              {{t "pages.team.members.actions.edit-role"}}
-            </Dropdown::Item>
-            <Dropdown::Item @onClick={{fn @onRemoveMemberButtonClicked @member}}>
-              {{t "pages.team.members.actions.remove-membership"}}
-            </Dropdown::Item>
-          {{/if}}
-          {{#if this.shouldDisplayLeaveCertificationCenterOption}}
-            <Dropdown::Item @onClick={{@onLeaveCertificationCenterButtonClicked}}>
-              {{t "pages.team.members.actions.leave-certification-center"}}
-            </Dropdown::Item>
-          {{/if}}
-        </Dropdown::IconTrigger>
-      {{/if}}
-      {{#if this.isEditionMode}}
-        <div class="members-list-item__managing-role">
-          <PixButton
-            id="save-certification-center-role"
-            @triggerAction={{fn this.updateMember @member}}
-            @size="small"
-            aria-label={{t "pages.team.members.actions.save"}}
-          >
-            {{t "pages.team.members.actions.save"}}
-          </PixButton>
-          <PixIconButton
-            @iconName="close"
-            id="cancel-update-certification-center-role"
-            @ariaLabel="{{t 'common.actions.cancel'}}"
-            @triggerAction={{this.cancelUpdateRoleOfMember}}
-            @withBackground={{false}}
-          />
-        </div>
-      {{/if}}
-    </td>
-  {{/if}}
-  {{#if @shouldDisplayRefererColumn}}
-    <td>
-      {{#if @member.isReferer}}
-        <div class="members-list-item__container">
-          <PixTag class="members-list-item__tag" @color="primary">
-            {{t "pages.team.pix-referer"}}
-          </PixTag>
-          <PixTooltip class="members-list-item__tooltip" @isWide="true" @position="left">
-            <:triggerElement>
-              <span tabindex="0">
-                <PixIcon
-                  @name="info"
-                  @plainIcon={{true}}
-                  @ariaHidden={{true}}
-                  class="members-list-item__tooltip-icon"
-                />
-              </span>
-            </:triggerElement>
-            <:tooltip>
-              {{t "pages.team.pix-referer-tooltip"}}
-            </:tooltip>
-          </PixTooltip>
-        </div>
-      {{/if}}
-    </td>
-  {{/if}}
-</tr>
