+++ certif/app/components/members-list.hbs
@@ -1,34 +1,11 @@
 <div role="tabpanel">
-  <div class="panel">
-    <div class="table content-text--small">
-      <table>
-        <thead>
-          <tr>
-            <Table::Header>{{t "common.labels.candidate.lastname"}}</Table::Header>
-            <Table::Header>{{t "common.labels.candidate.firstname"}}</Table::Header>
-            <Table::Header>{{t "common.labels.candidate.role"}}</Table::Header>
-            {{#if this.shouldDisplayManagingColumn}}
-              <Table::Header>{{t "pages.team.table-headers.actions"}}</Table::Header>
-            {{/if}}
-            {{#if this.shouldDisplayRefererColumn}}
-              <Table::Header>{{t "pages.team.referer"}}</Table::Header>
-            {{/if}}
-          </tr>
-        </thead>
-        <tbody>
-          {{#each @members as |member|}}
-            <MembersListItem
-              @member={{member}}
-              @shouldDisplayRefererColumn={{this.shouldDisplayRefererColumn}}
-              @isMultipleAdminsAvailable={{this.isMultipleAdminsAvailable}}
-              @onLeaveCertificationCenterButtonClicked={{this.openLeaveCertificationCenterModal}}
-              @onRemoveMemberButtonClicked={{this.openRemoveMemberModal}}
-            />
-          {{/each}}
-        </tbody>
-      </table>
-    </div>
-  </div>
+  <MembersTable
+    @members={{@members}}
+    @hasCleaHabilitation={{@hasCleaHabilitation}}
+    @onLeaveCertificationCenterButtonClicked={{this.openLeaveCertificationCenterModal}}
+    @onRemoveMemberButtonClicked={{this.openRemoveMemberModal}}
+    @onChangeMemberRoleButtonClicked={{this.openChangeMemberRoleModal}}
+  />
 </div>
 
 <Team::Modal::LeaveCertificationCenter
@@ -37,10 +14,18 @@
   @onClose={{this.closeLeaveCertificationCenterModal}}
   @onSubmit={{this.leaveCertificationCenter}}
 />
+
 <Team::Modal::RemoveMemberModal
   @firstName={{this.removingMember.firstName}}
   @lastName={{this.removingMember.lastName}}
   @isOpen={{this.isRemoveMemberModalOpen}}
   @onSubmit={{this.removeMember}}
   @onClose={{this.closeRemoveMemberModal}}
+/>
+
+<Team::Modal::ChangeMemberRoleModal
+  @member={{this.member}}
+  @isOpen={{this.isChangeMemberRoleModalOpen}}
+  @onSubmit={{this.changeMemberRole}}
+  @onClose={{this.closeChangeMemberRoleModal}}
 />
