+++ certif/app/styles/app.scss
@@ -51,6 +51,7 @@
 @use 'components/session-supervising/handle-live-alert-modal' as *;
 @use 'components/session-supervising/live-alert-handled-modal' as *;
 @use 'components/team/invitations-list' as *;
+@use 'components/team/change-member-role-modal' as *;
 @use 'components/dropdown' as *;
 @use 'flatpickr' as *;
