+++ certif/app/styles/components/team/change-member-role-modal.scss
@@ -1,0 +1,7 @@
+.change-member-role-modal__select-role {
+  padding: var(--pix-spacing-4x) 0;
+
+  button {
+    width: fit-content;
+  }
+}
