+++ certif/translations/fr.json
@@ -958,18 +958,21 @@
         "actions": {
           "edit-role": "Modifier le rôle",
           "leave-certification-center": "Quitter cet espace Pix Certif",
-          "manage": "Gérer",
           "remove-membership": "Supprimer",
-          "save": "Enregistrer",
-          "select-role": {
-            "label": "Sélectionner un rôle",
-            "options": {
-              "admin": "Administrateur",
-              "member": "Membre"
-            }
-          }
+          "save": "Enregistrer"
         },
         "modals": {
+          "change-member-role": {
+            "title": "Modifier le rôle",
+            "information": "Vous vous apprêtez à changer le rôle de {firstName} {lastName}.",
+            "select-role": {
+              "label": "Sélectionner un rôle",
+              "options": {
+                "admin": "Administrateur",
+                "member": "Membre"
+              }
+            }
+          },
           "leave-certification-center": {
             "title": "Quitter cet espace Pix Certif",
             "message": "Êtes-vous sûr de vouloir quitter cet espace Pix Certif ? Vous n’aurez plus accès à <strong>{certificationCenterName}</strong>.<br/><br/>Vos sessions restent accessibles au reste de l'équipe.<br/><br/>Vos accès à d’autres espaces Pix Certif ne seront pas impactés.<br/><br/>Vous allez être déconnecté."
@@ -1022,6 +1025,9 @@
         "description": "Le Groupement d’intérêt public, sis 21-23 rue des Ardennes 75019 Paris PIX met en œuvre en tant que responsable de traitement un traitement de données à caractère personnel portant sur votre adresse email professionnel de référent Pix.<br />La finalité de ce traitement est la communication des informations nécessaires aux attributions du référent Pix au sein du centre de certification agréé. Le traitement de cette données repose sur l’exécution de la convention d’agrément de votre centre de certification.<br />Les données sont conservées pendant la durée de la convention d’agrément en vigueur entre le GIP Pix et le centre habilité auquel vous êtes rattaché.<br />Vous disposez de droits personnels, droits d'accès, de rectification et d’effacement des données, le droit de limitation ou d’opposition au traitement pour motif légitime et le droit d’introduire une réclamation auprès d’une autorité de contrôle ainsi que le droit de définir des directives relatives à la conservation, à l'effacement et à la communication de vos données à caractère personnel en cas de décès que vous pouvez exercer par courriel auprès de notre DPD en adressant votre demande à l’adresse électronique suivante : dpd@pix.fr .<br />Vous disposez de la possibilité d’introduire une réclamation auprès de Commission nationale Informatique et libertés « CNIL » à l’adresse suivante : 3 Place de Fontenoy – TSA 80715 – 75334 Paris Cedex 07.",
         "empty-option": "--Sélectionner--",
         "label": "Sélectionner le référent CléA Numérique"
+      },
+      "table": {
+        "caption": "Tableau listant les membres du centre de certification. Il comporte le nom et prénom du membre ainsi que le rôle. Si l'utilisateur connecté possède un rôle administrateur, une colonne action permet de modifier le rôle d'un membre, de supprimer un membre de l'espace ou encore de quitter lui-même cet espace de certification. Si le centre possède une habilitation CléA,le tableau comportera le référent."
       },
       "table-headers": {
         "actions": "Actions"
